[SyncFS] LocalChangeProcessorDelegate::Delete should not hit DCHECK on missing metadata.
BUG=248422
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/16633005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@205404 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: d918732646632e7f820fb038ced5882088309bc3

+++ chrome/browser/sync_file_system/drive/local_change_processor_delegate.cc
@@ -248,7 +248,10 @@ void LocalChangeProcessorDelegate::DidUploadExistingFile(
 
 void LocalChangeProcessorDelegate::Delete(
     const SyncStatusCallback& callback) {
-  DCHECK(has_drive_metadata_);
+  if (!has_drive_metadata_) {
+    DidDelete(callback, google_apis::HTTP_NOT_FOUND);
+    return;
+  }
   api_util()->DeleteFile(
       drive_metadata_.resource_id(),
       drive_metadata_.md5_checksum(),
